Switch audio input IPC to mojo.
Mojo implementations for audio input stream creation and control are now
implemented. This CL changes the AudioDeviceFactory to use the new code.
The old code will be removed in a followup CL.

This CL adds an "AudioInputIPCFactory", and renames the current
"AudioIPCFactory" to "AudioOutputIPCFactory" to avoid confusion.

Bug: 653871

+++ content/browser/renderer_host/media/render_frame_audio_input_stream_factory.h
@@ -63,13 +63,13 @@ class CONTENT_EXPORT RenderFrameAudioInputStreamFactory
 
   // mojom::RendererAudioInputStreamFactory implementation.
   void CreateStream(mojom::RendererAudioInputStreamFactoryClientPtr client,
-                    int32_t session_id,
+                    int64_t session_id,
                     const media::AudioParameters& audio_params,
                     bool automatic_gain_control,
                     uint32_t shared_memory_count) override;
 
   void DoCreateStream(mojom::RendererAudioInputStreamFactoryClientPtr client,
-                      int session_id,
+                      int64_t session_id,
                       const media::AudioParameters& audio_params,
                       bool automatic_gain_control,
                       uint32_t shared_memory_count,
